requirements.txt
aiogram
yt-dlp
shazamio
aiosqlite
python-dotenv
ffmpeg-python